refactor: move picker
