+++ docs/examples/debug.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import '../../assets/index.less';
 import type { Locale, PickerRef } from '../../src/interface';
-import RangePicker from '../../src/NewPicker/PickerInput/RangePicker';
-import SinglePicker from '../../src/NewPicker/PickerInput/SinglePicker';
+import RangePicker from '../../src/PickerInput/RangePicker';
+import SinglePicker from '../../src/PickerInput/SinglePicker';
 import PickerPanel from '../../src/PickerPanel';
 
 import dayjs, { type Dayjs } from 'dayjs';
