+++ src/NewPicker/PickerInput/Selector/SingleSelector/index.tsx
@@ -2,8 +2,8 @@ import classNames from 'classnames';
 import * as React from 'react';
 import { isSame } from '../../../../utils/dateUtil';
 import type { InternalMode, SelectorProps, SelectorRef } from '../../../../interface';
-import PickerContext from '../../context';
-import type { PickerProps } from '../../SinglePicker';
+import PickerContext from '../../../../PickerInput/context';
+import type { PickerProps } from '../../../../PickerInput/SinglePicker';
 import useInputProps from '../hooks/useInputProps';
 import useRootProps from '../hooks/useRootProps';
 import Icon, { ClearIcon } from '../Icon';
