+++ src/NewPicker/PickerInput/hooks/useFilledProps.ts
@@ -4,7 +4,7 @@ import { toArray } from '../../../utils/miscUtil';
 import { fillLocale } from '../../../hooks/useLocale';
 import { getTimeConfig } from '../../../hooks/useTimeConfig';
 import type { FormatType, InternalMode } from '../../../interface';
-import type { RangePickerProps } from '../RangePicker';
+import type { RangePickerProps } from '../../../PickerInput/RangePicker';
 import { fillClearIcon } from '../Selector/hooks/useClearIcon';
 import useDisabledBoundary from './useDisabledBoundary';
 import { useFieldFormat } from './useFieldFormat';
