+++ src/PickerInput/context.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import type { GenerateConfig } from '../../generate';
-import type { Components, Locale } from '../../interface';
+import type { GenerateConfig } from '../generate';
+import type { Components, Locale } from '../interface';
 
 export interface PickerContextProps<DateType = any> {
   prefixCls: string;
